install_grub2(): execute update-grub *after* grub-install.
If files in /boot/grub are missing that are installed by grub-install
the syntax check of grub.cfg fails (due to broken load_video() config),
resulting in a grub.cfg.new file which would *basically* work and
a grub.cfg that's shipped by the live system and definitely does not work.

Executing update-grub after grub-install fixes this problem.